golang decoder

+++ golang/encoder.go
@@ -5,15 +5,15 @@ import (
 )
 
 // Encode takes a decimal value and transforms it to its base-62 string representation
-func Encode(value int) (encoded string, err error) {
+func Encode(value uint) (encoded string, err error) {
 	for value > 0 {
-		v, idx, e := utls.EuclideanDivision(value, 62)
+		v, idx, e := utls.EuclideanDivision(int(value), 62)
 		if e != nil {
 			err = e
 			return
 		}
 		encoded = string(CHARSET[idx]) + encoded
-		value = v
+		value = uint(v)
 	}
 	return
 }
